[Simpler Update] Version 2.2
Made a few mistakes with calculations and variables which have now been fixed, your damage or healing will now reset properly when switching between values but percentage will stay until you press clear. Also there is now a descriptions box so you can put card titles in the box next to the damage value to keep track. YES cards will be coming but in a separate window so it isn't as bulky as older versions, there will be a button to open the card menu which you can move around your desktop for easy access, it will be horizontal so you can align it above or below your wiz window.