[Simpler Version 2.4.1] Patch
Corrected more of the bugs with numbers, crit was removed as it wasn't working correctly with the check box, it will be reintroduced at a later stage. The ui has also had a nice red update to keep things fresh between versions. Descriptions was removed as we usually know what spells we are using, but some more features along that topic will be added in the future.